delete updated

+++ server/controllers/clerkWebhooks.js
@@ -32,30 +32,46 @@ const clerkWebhooks = async (req, res) => {
         await whook.verify(JSON.stringify(req.body), headers);
 
         // Getting Data from request body
-        const {data, type} = req.body
+        const { data, type } = req.body;
 
         console.log("Processing webhook event:", { type, userId: data?.id });
 
-        const userData = {
-            _id: data.id,
-            email: data.email_addresses[0].email_address,
-            username: data.first_name + " " + data.last_name,
-            image: data.image_url,
+        if (!type || !data || !data.id) {
+            throw new Error("Invalid webhook payload: missing type or data.id");
         }
 
         // Switch cases for different events
         switch (type) {
-            case "user.created":{
-                console.log("Creating user:", userData);
-                await User.create(userData);
+            case "user.created": {
+                const primaryEmail = data.email_addresses?.[0]?.email_address;
+                const createdUser = {
+                    _id: data.id,
+                    email: primaryEmail,
+                    username: `${data.first_name || ""} ${data.last_name || ""}`.trim(),
+                    image: data.image_url,
+                };
+                console.log("Creating user:", { _id: createdUser._id, email: createdUser.email });
+                await User.create(createdUser);
                 break;
             }
-            case "user.updated":{
-                console.log("Updating user:", userData);
-                await User.findByIdAndUpdate(data.id ,userData);
+            case "user.updated": {
+                const update = {};
+                if (data.first_name || data.last_name) {
+                    update.username = `${data.first_name || ""} ${data.last_name || ""}`.trim();
+                }
+                if (data.image_url) {
+                    update.image = data.image_url;
+                }
+                if (data.email_addresses?.[0]?.email_address) {
+                    update.email = data.email_addresses[0].email_address;
+                }
+                console.log("Updating user:", { _id: data.id, updateFields: Object.keys(update) });
+                if (Object.keys(update).length > 0) {
+                    await User.findByIdAndUpdate(data.id, update, { new: true });
+                }
                 break;
             }
-            case "user.deleted":{
+            case "user.deleted": {
                 console.log("Deleting user:", data.id);
                 await User.findByIdAndDelete(data.id);
                 break;
